requirements.txt
# requirements.txt

# For web browser automation
selenium
webdriver-manager
pandas
openpyxl
python-docx
docxtpl
Flask
ocrmypdf
celery
redis
pypdf


# For managing environment variables (secrets)
python-dotenv